+++ utils/unwind/BUILD.bazel
@@ -4,7 +4,7 @@ load("//.bazel:defs.bzl", "go_test")
 go_library(
     name = "unwind",
     srcs = ["unwind.go"],
-    importpath = "github.com/ava-labs/avalanchego/vms/saevm/unwind",
+    importpath = "github.com/ava-labs/avalanchego/utils/unwind",
     visibility = ["//visibility:public"],
 )
 
@@ -16,6 +16,7 @@ go_test(
     ],
     embed = [":unwind"],
     deps = [
+        "@com_github_ava_labs_avalanchego//vms/saevm/unwind",
         "@com_github_google_go_cmp//cmp",
         "@com_github_google_go_cmp//cmp/cmpopts",
         "@com_github_stretchr_testify//assert",
