+++ vms/saevm/sae/BUILD.bazel
@@ -35,6 +35,7 @@ go_library(
         "//utils/bloom",
         "//utils/constants",
         "//utils/logging",
+        "//utils/unwind",
         "//version",
         "//vms/components/gas",
         "//vms/saevm/adaptor",
@@ -48,7 +49,6 @@ go_library(
         "//vms/saevm/saexec",
         "//vms/saevm/txgossip",
         "//vms/saevm/types",
-        "//vms/saevm/unwind",
         "//vms/saevm/worstcase",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//core",
